% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageFormat.R
\name{ImageExtract}
\alias{ImageExtract}
\title{Function to format Weeview 3d video images for use in SEBASTES stereo-analysis software}
\usage{
ImageExtract(pathFrom = NA, pathTo = NA)
}
\arguments{
\item{pathFrom}{This is the folder location where the movies to be converted are located}

\item{pathTo}{This is the folder location where you want the formated images and metadata to go}
}
\description{
This function converts video collected by a Weeview 3d camera to paired images and metadata that can be used for stereo-image analysis
in SEBASTES software. The function prompts the user to point to a folder where the Weeview movies are located that need
to be converted. Then it will print the list of movies in that folder. Next it will prompt the user for a location to place
the converted images and the metadata required by Sebastes. It will then sample the movie for individual frames at 1 second intervals,
do the image conversion and produce the required metadata. #' It is important to note that the metadata generated by this function is
not informative about the images. The function can also #' be run from the command line using the second example where the from and to
file folders are specified.
}
\examples{
Example 1: User is prompted for file locations
ImageExtract()

Example 2: Command line specification of file locations
ImageExtract("D:/Weeview/MOVIE","C:/Users/rooperc/Desktop/WeeViewTest2")
}
\keyword{3d}
\keyword{SEBASTES,}
\keyword{Stereo}
\keyword{Weeview}
\keyword{camera}
\keyword{imagery,}
