% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageFormat.R
\name{ImageFormat}
\alias{ImageFormat}
\title{Function to format Weeview 3d camera images for use in SEBASTES stereo-analysis software}
\usage{
ImageFormat(pathFrom = NA, pathTo = NA)
}
\arguments{
\item{pathFrom}{This is the folder location where the photos to be converted are located}

\item{pathTo}{This is the folder location where you want the formated images and metadata to go}
}
\description{
This function converts images collected by a Weeview 3d camera to paired images and metadata that can be used for stereo-image analysis
in SEBASTES software. The function prompts the user to point to a folder where the Weeview photos are located that need
to be converted. Then it will print the list of images in that folder. Next it will prompt the user for a location to place
the converted images and the metadata required by Sebastes. It will then do the image conversion and produce the required metadata.
It is important to note that the metadata generated by this function is not informative about the images. The function can also
be run from the command line using the second example where the from and to file folders are specified.
}
\examples{
Example 1: User is prompted for file locations
ImageFormat()

Example 2: Command line specification of file locations
ImageFormat("D:/Weeview/PHOTO","C:/Users/rooperc/Desktop/WeeViewTest")
}
\keyword{3d}
\keyword{SEBASTES,}
\keyword{Stereo}
\keyword{Weeview}
\keyword{camera}
\keyword{imagery,}
